selemium test tested and done

+++ TYPEit Web App/Main Folder/flask_app/tests/frontend.py
@@ -17,12 +17,6 @@ from selenium.webdriver.common.by import By
 
 basedir = os.path.abspath(os.path.dirname(__file__))
 
-#driver = webdriver.Firefox(executable_path=os.path.join(basedir,'geckodriver'))
-#driver.get('http://localhost:5000/')
-#time.sleep(1)
-#driver.find_element_by_name('username').send_keys('Tom')
-#driver.find_element_by_name('password').send_keys('pw')
-
 app = Flask(__name__)
 app=create_app(config_file='test_settings.py')
 app.testing = True
@@ -30,33 +24,28 @@ app.test_client()
 
 # --------------------------------- Global Variables ---------------------------------------- #
 
-test_admin_user = 'admin'
-test_admin_password='admin'
-
-test_incorrect_user='luke'
-test_incorrecct_password='skywalker123'
-
-test_student_user = 'akhil'
-test_student_password = 'akhil'
 
 class TestBase(unittest.TestCase):
     driver = None
     def setUp(self):
-        with app.app_context():
-            db.session.commit()
-            db.drop_all()
-            db.create_all()
-            self.driver = webdriver.Firefox(executable_path=os.path.join(basedir,'geckodriver'))
-            self.driver.implicitly_wait(30)
-            self.base_url = 'http://localhost:5000/'
-            admin = User(username="admin", password=generate_password_hash("admin"), admin=True, student=False, enabled=True)
-            student1 = User(username='akhil',password=generate_password_hash('akhil'), admin=False, student=True, enabled=True)
-            db.session.add(admin)
-            db.session.add(student1)
-            db.session.commit()
-            #db.drop_all()
-            self.driver.get('http://localhost:5000/login')
-            self.driver.maximize_window()
+        self.driver = webdriver.Firefox(executable_path=os.path.join(basedir,'geckodriver'))
+        if not self.driver:
+            self.skipTest
+        else:
+            with app.app_context():
+                db.session.commit()
+                db.drop_all()
+                db.create_all()
+                self.driver.implicitly_wait(30)
+                self.base_url = 'http://localhost:5000/'
+                admin = User(username="admin", password=generate_password_hash("admin"), admin=True, student=False, enabled=True)
+                student1 = User(username='akhil',password=generate_password_hash('akhil'), admin=False, student=True, enabled=True)
+                db.session.add(admin)
+                db.session.add(student1)
+                db.session.commit()
+                #db.drop_all()
+                self.driver.get('http://localhost:5000/login')
+                self.driver.maximize_window()
            
 
     def tearDown(self):
@@ -70,10 +59,11 @@ class TestBase(unittest.TestCase):
 
 #New user should be able to successfully create an account
 #Once, the user is registered, the browser should redirect to the login page
-
-
+    
+    
 class TestRegisteration(TestBase):
-    '''
+    
+
     def test_invalid_register(self):
 
         button_field = self.driver.find_element_by_id('log-in-register-btn').click()
@@ -84,15 +74,15 @@ class TestRegisteration(TestBase):
         username_taken_message = self.driver.find_element_by_class_name('alert').text
         self.assertEqual(username_taken_message, 'This username is taken, try again.')
 
-
+    
     def test_register(self):
         button_field = self.driver.find_element_by_id('log-in-register-btn').click()
         time.sleep(1)
-        user_field = self.driver.find_element_by_id('username-input').send_keys('Ridkkk')
+        user_field = self.driver.find_element_by_id('username-input').send_keys('lukeee')
         password_field = self.driver.find_element_by_id('password-input').send_keys('Morty1234')
         button_field = self.driver.find_element_by_id('register-btn').click()
         time.sleep(1)
-        user_field = self.driver.find_element_by_id('username-input').send_keys('Ridkkk')
+        user_field = self.driver.find_element_by_id('username-input').send_keys('lukeee')
         password_field = self.driver.find_element_by_id('password-input').send_keys('Morty1234')
         button_field = self.driver.find_element_by_id('log-in-btn').click()
         time.sleep(3)
@@ -137,7 +127,7 @@ class TestLogin(TestBase):
         #admin access
         #self.driver.get(self.base_url+'/select-topics')
         time.sleep(2)
-        greetings = self.driver.find_element_by_id('Welcome').text
+        greetings = self.driver.find_element_by_id('Welcome').get_attribute('innerHTML')
         self.assertIn(greetings, 'Welcome, akhil!')
         #self.driver.find_element_by_partial_link_text("My").click()
         #time.sleep(5)
@@ -146,13 +136,16 @@ class TestLogin(TestBase):
         username_taken_message = self.driver.find_element_by_class_name('alert').text
         self.assertEqual(username_taken_message, 'You are now logged out')
     
-    '''
+    
 
 #admin disables a topic, student cannot click on that quix
 
 #disable user 
 
+'''
+
     def test_disable_student(self):
+         
         time.sleep(2)
         user = self.driver.find_element_by_id('username-input').send_keys('admin')
         password = self.driver.find_element_by_id('password-input').send_keys('admin')
@@ -161,110 +154,135 @@ class TestLogin(TestBase):
         time.sleep(0.5)
         self.driver.find_element_by_partial_link_text("Manage").click()
         time.sleep(2)
-        user = User.query.filter_by(username='akhil').first()
+        user = User.query.filter_by(username='akhil').all()
+        self.driver.find_element_by_class_name('student-switch-label').click()
+        self.driver.find_element_by_partial_link_text("Log").click()
+        username_taken_message = self.driver.find_element_by_class_name('alert').text
+        time.sleep(3)
+        self.assertEqual(username_taken_message, 'You are now logged out')
+        time.sleep(6)
         #if user.enabled is False:
         #    self.driver.find_element_by_class_name('student-switch-label').click()
-        if user.enabled is True:
-            self.driver.find_element_by_class_name('student-switch-label').click()
-            self.driver.find_element_by_partial_link_text("Log").click()
-            username_disabled_message = self.driver.find_element_by_class_name('alert').text
-            self.assertEqual(username_taken_message, 'You are now logged out')
-            time.sleep(3)
-            user = self.driver.find_element_by_id('username-input').send_keys('akhil')
-            password = self.driver.find_element_by_id('password-input').send_keys('akhil')
-            time.sleep(1)
-            button = self.driver.find_element_by_id('log-in-btn').click()
-            username_disabled_message = self.driver.find_element_by_class_name('alert').text
-            self.assertEqual(username_taken_message, 'User disabled, contact Admin.')
-      
+        user = self.driver.find_element_by_id('username-input').send_keys('akhil')
+        password = self.driver.find_element_by_id('password-input').send_keys('akhil')
+        button = self.driver.find_element_by_id('log-in-btn').click()
+        disabed = self.
+        if user.enabled is False: 
+            username_disabled_message = self.driver.find_element_by_id('alert-message').text
+            self.assertEqual(username_disabled_message, 'User disabled, contact Admin.')
+            time.sleep(2)
+        else: 
+            self.driver.find_element_by_partial_link_text("My").click()
+            time.sleep(2)
 
 
 
     '''
 
-#check if student can enter a quiz
+class TestQuiz(TestBase):
 
-    def test_quiz(self):
+    def test_play_quiz(self):
         time.sleep(2)
         user = self.driver.find_element_by_id('username-input').send_keys('akhil')
         password = self.driver.find_element_by_id('password-input').send_keys('akhil')
         time.sleep(1)
         button = self.driver.find_element_by_id('log-in-btn').click()
-        self.driver.find_element_by_id("English").click()
+        time.sleep(0.5)
+        button = self.driver.find_element_by_id('English').click()
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('i')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('m')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('p')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('r')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('o')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('m')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('p')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('t')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('u')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('h')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('a')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('i')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('k')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('u')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('d')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('e')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('x')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('t')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('e')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('r')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('i')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('t')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('y')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('p')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('h')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('o')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('b')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('i')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('a')
+        time.sleep(0.5)
+        button = self.driver.find_element_by_id('skip-btn').click()
+        time.sleep(0.5)
+        button = self.driver.find_element_by_id('skip-btn').click()
+        time.sleep(1)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('d')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('e')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('r')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('v')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('i')
+        time.sleep(1)
+        button = self.driver.find_element_by_id('skip-btn').click()
+        time.sleep(1)
+        button = self.driver.find_element_by_id('skip-btn').click()
+        time.sleep(1)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('t')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('o')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('r')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('q')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('u')
+        time.sleep(0.5)
+        self.driver.find_element_by_id('hint-and-answer-input').send_keys('e')
+        time.sleep(1)
+        button = self.driver.find_element_by_id('skip-btn').click()
+        time.sleep(3)
+        button = self.driver.find_element_by_id('play-again-btn').click()
         time.sleep(2)
-        self.driver.implicitly_wait(5)
-        self.driver.find_element_by_id("hint-and-answer-input").send_keys('impromptu')
-        self.driver.implicitly_wait(5)
-        time.sleep(4)
-
-
-    '''
-
-''' 
-    
-class TestLogin(TestBase):
-
-
-    def test_login(self):
-        time.sleep(1)
-        #admin login 
-        user_field = self.driver.find_element_by_id('username-input').send_keys('admin')
-        password_field = self.driver.find_element_by_id('password-input').send_keys('admin')
-        button_field = self.driver.find_element_by_id('log-in-btn').click()
-
-        #Application should redirect to the Select Topics Page
-        self.driver.get(self.base_url+'/select-topics')
-        welcome_message = self.driver.find_element_by_id('Welcome').get_attribute('innerHTML')
-        self.assertIn(welcome_message, 'Welcome, Admin!')
-
-
-
-
-    def test_access_denied(self):
-        user = self.driver.find_element_by_id('username-input').send_keys('akhil')
-        password = self.driver.find_element_by_id('password-input').send_keys('akhil')
-        button = self.driver.find_element_by_id('log-in-btn').click()
-
-        url_tag = self.driver.get(self.base_url+'/select-topics')
-
-        self.driver.get(url_tag)
-        error = self.driver.find_element_by_css_selector('h1').text
-        self.assertIn('You do not have admin access. Go back!', error)
-
-        
-
-
-
-
-
-
-    #incorrect login details, user is not in database
-    def test_incorrect_login(self):
-     
-        time.sleep(1)
-        
-        user_field = self.driver.find_element_by_id('username-input').send_keys('abfnd')
-        #self.driver.find_element_by_name('username-input').clear()
-        password_field = self.driver.find_element_by_id('password-input').send_keys('asbhdjn')
-        #self.driver.find_element_by_name('password-input').clear()
-        button_field = self.driver.find_element_by_id('log-in-btn').click()
-  
-     
-    
-
-    def test_correct_login(self):
-        
-        time.sleep(1)
-        user_field = self.driver.find_element_by_id('username-input').send_keys('akhil')
-        password_field = self.driver.find_element_by_id('password-input').send_keys('akhil')
-        button_field = self.driver.find_element_by_id('log-in-btn').click()
-
-
-
-
-
-
-'''
+        self.driver.find_element_by_partial_link_text("My").click()
+        time.sleep(2)
 
 
 
@@ -274,53 +292,8 @@ if __name__=='__main__':
   unittest.main(verbosity=2)
         
         
-       # tester = app.test_client()
-       # response = tester.get('/', content_type='html/text')
-        #self.assertEqual(response.status_code, 302)
-        #admin = User.query.get(1)
-        #self.assertEquals(admin.username, 'Test')
-        #userr = User.query.filter_by(username='admin')
-        #driver = self.driver 
-       # self.driver.get('http://localhost:5000/')
-
-       # user = driver.find_element_by_name('username').send_keys('admin')
-       # passwd = driver.find_element_by_name('password').send_keys('admin')
-       # assertEqual(userr, user)
-
-
-
-
-'''
-class Login(object):
-
-    def __init__()
-
-
-'''
 
 
 
 
 
-
-
-
-
-
-
-
-
-
-
-
-#launch the browser fireforx, naviagte to url and look for element (textbox by name q) then type in software testing in there
-
-
-#ccreate driver object - instalise the webbroswer with firefox
-#basedir = os.path.abspath(os.path.dirname(__file__))
-
-#navigate to the following URL
-#driver.get('http://localhost:5000/')
-
-
-#driver.find_element_by_name('username').send_keys('software testing')
